Project_Live/requirements.txt
kiteconnect==5.0.1
streamlit
pandas
mysql-connector-python
python-dotenv
